Move StandardComponents
Components refer to all parts of strata-measure
