+++ examples/src/main/java/com/opengamma/strata/examples/finance/CalibrationEur3CheckExample.java
@@ -6,7 +6,7 @@
 package com.opengamma.strata.examples.finance;
 
 import static com.opengamma.strata.collect.Guavate.toImmutableList;
-import static com.opengamma.strata.measure.calc.StandardComponents.marketDataFactory;
+import static com.opengamma.strata.measure.StandardComponents.marketDataFactory;
 import static java.util.stream.Collectors.toMap;
 
 import java.time.LocalDate;
@@ -40,7 +40,7 @@ import com.opengamma.strata.market.curve.CurveGroupName;
 import com.opengamma.strata.market.observable.QuoteId;
 import com.opengamma.strata.market.product.deposit.IborFixingDepositCurveNode;
 import com.opengamma.strata.measure.Measures;
-import com.opengamma.strata.measure.calc.StandardComponents;
+import com.opengamma.strata.measure.StandardComponents;
 import com.opengamma.strata.measure.rate.RatesMarketDataLookup;
 import com.opengamma.strata.product.Trade;
