+++ examples/src/test/java/com/opengamma/strata/examples/finance/credit/harness/TestCalculator.java
@@ -26,7 +26,7 @@ import com.opengamma.strata.data.scenario.ScenarioMarketData;
 import com.opengamma.strata.examples.marketdata.ExampleMarketData;
 import com.opengamma.strata.examples.marketdata.ExampleMarketDataBuilder;
 import com.opengamma.strata.market.param.CurrencyParameterSensitivities;
-import com.opengamma.strata.measure.calc.StandardComponents;
+import com.opengamma.strata.measure.StandardComponents;
 import com.opengamma.strata.product.Trade;
 
 public class TestCalculator implements Calculator {
