--- modules/measure/src/main/java/com/opengamma/strata/measure/calc/package-info.java
@@ -1,11 +1,0 @@
-/**
- * Copyright (C) 2015 - present by OpenGamma Inc. and the OpenGamma group of companies
- *
- * Please see distribution for license.
- */
-
-/**
- * Integration code that allows strata-calc to operate on the built-in asset classes.
- */
-package com.opengamma.strata.measure.calc;
-
